New stuff - modified bash_profile - modified setup.sh - added a script to configure xrandr on 2560x1080 display

+++ setup.sh
@@ -1,12 +1,20 @@
 #!/usr/bin/env bash
 
-# Backup of existend files
-mv ~/.bash_profile ~/.bash_profile.orig
-mv ~/.bashrc ~/.bashrc.orig
-mv ~/.vimrc ~/.vimrc.orig
-mv ~/.gitconfig ~/.gitconfig.orig
+# Backup of existent files
+if [ -f "~/.bash_profile" ]; then
+    mv ~/.bash_profile ~/.bash_profile.orig
+fi
+if [ -f "~/.bashrc" ]; then
+    mv ~/.bashrc ~/.bashrc.orig
+fi
+if [ -f "~/.vimrc" ]; then
+    mv ~/.vimrc ~/.vimrc.orig
+fi
+if [ -f "~/.gitconfig" ]; then
+    mv ~/.gitconfig ~/.gitconfig.orig
+fi
 
-# Link to repo's files
+# Copy repo files
 cp ./bash_profile ~/.bash_profile
 cp ./bashrc ~/.bashrc
 cp ./vimrc ~/.vimrc
